+++ src/allotropy/allotrope/converter.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from collections.abc import Mapping
+from collections.abc import Mapping, Sequence
 from dataclasses import asdict, fields, is_dataclass, make_dataclass
 from types import UnionType
 from typing import Any, Callable, cast, get_args, get_origin, TypeVar, Union
@@ -93,9 +93,17 @@ ModelClass = TypeVar("ModelClass")
 
 
 def add_custom_information_document(
-    model: ModelClass, custom_info_doc: dict[str, Any]
+    model: ModelClass, custom_info_doc: Any
 ) -> ModelClass:
-    model.custom_information_document = structure_custom_information_document(custom_info_doc, "custom information document")  # type: ignore
+
+    if isinstance(custom_info_doc, dict):
+        custom_info_doc = structure_custom_information_document(
+            custom_info_doc, "custom information document"
+        )
+    if not is_dataclass(custom_info_doc):
+        msg = "Invalid custom_info_doc"
+        raise ValueError(msg)
+    model.custom_information_document = custom_info_doc  # type: ignore
     return model
 
 
@@ -208,6 +216,22 @@ def structure_custom_information_document(val: dict[str, Any], name: str) -> Any
     )
 
 
+# Special should_omit check for allowing an empty value for 'value' keys, controlled by should_allow_empty_value_field
+def should_omit_allow_empty_value_field(k: str, v: Any) -> bool:
+    return v is None and k != "value"
+
+
+def unstructure_custom_information_document(model: Any) -> dict[str, Any]:
+    def dict_factory(kv_pairs: Sequence[tuple[str, Any]]) -> dict[str, Any]:
+        return {
+            _convert_model_key_to_dict_key(key): value
+            for key, value in kv_pairs
+            if not should_omit_allow_empty_value_field(key, value)
+        }
+
+    return asdict(model, dict_factory=dict_factory)
+
+
 def register_dataclass_hooks(converter: Converter) -> None:
     def dataclass_structure_fn(cls: Any) -> Callable[[Any, Any], Any | None]:
         structure_fn = make_dict_structure_fn(
@@ -277,10 +301,6 @@ def register_unstructure_hooks(converter: Converter) -> None:
     def should_omit(_: str, v: Any) -> bool:
         return v is None
 
-    # Special should_omit check for allowing an empty value for 'value' keys, controlled by should_allow_empty_value_field
-    def should_omit_allow_empty_value_field(k: str, v: Any) -> bool:
-        return v is None and k != "value"
-
     unstructure_fn_cache = {}
 
     def unstructure_dataclass_fn(
@@ -297,7 +317,9 @@ def register_unstructure_hooks(converter: Converter) -> None:
                 if not should_omit(k, v)
             }
             if hasattr(obj, "custom_information_document"):
-                dataclass_dict["custom information document"] = asdict(
+                dataclass_dict[
+                    "custom information document"
+                ] = unstructure_custom_information_document(
                     obj.custom_information_document
                 )
             return dataclass_dict