+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added utility to add a custom information document to an ASM model
+- Added utility to add both dict and dataclass custom information document to an ASM model
 
 ### Fixed